ctx: drop zlib dep use tinf internally instead

+++ src/vt-audio.h
@@ -18,7 +18,6 @@
 
 #if CTX_SDL
 #include <SDL.h>
-#include <zlib.h>
 
 static int ydec (const void *srcp, void *dstp, int count)
 {
@@ -1388,7 +1387,7 @@ void vt_audio (VT *vt, const char *command)
     {
       case 'z':
     {
-      unsigned long actual_uncompressed_size = audio->frames * audio->bits/8 * audio->channels + 512;
+      unsigned int actual_uncompressed_size = audio->frames * audio->bits/8 * audio->channels + 512;
       unsigned char *data2 = malloc (actual_uncompressed_size);
       /* if a buf size is set (rather compression, but
        * this works first..) then */
